+++ src/lib.rs
@@ -18,6 +18,8 @@ pub const MU0_4PI: f64 = 1e-7;
 /// $$\mu_0 = 4\pi \cdot 10^{-7} H/m$$
 pub const MU0: f64 = 4.0 * PI * MU0_4PI;
 
+pub mod types;
+
 /// Expressions for various Biot-Savart sources
 pub mod sources;