+++ src/python.rs
@@ -6,16 +6,18 @@ use numpy::{
 };
 use pyo3::prelude::*;
 
-use crate::biotsavart_parallel;
-use crate::mesh;
-use crate::octree::{CurrentSources, HFieldSolver, Octree, point};
-use crate::vec3::Vec3;
+use crate::{
+    biotsavart_parallel, mesh,
+    octree::{CurrentSources, DipoleSources, HFieldSolver, Octree, point, tet_element},
+    types::{to_u32x4s, to_vec3s, to_vec3s_mut},
+    vec3::Vec3,
+};
 
 // ---
 // Helpers
 // ---
 
-// Transpose a row-major (N,3) PyArray2 to column vectors for SIMD operations in Rust
+// Transpose a row-major (N,3) Python array to column vectors for SIMD operations in Rust
 fn pyarray_to_3cols<T: Element + Copy>(arr: PyReadonlyArray2<T>) -> (Vec<T>, Vec<T>, Vec<T>) {
     let n = arr.shape()[0];
     let mut x: Vec<T> = Vec::with_capacity(n);
@@ -27,6 +29,17 @@ fn pyarray_to_3cols<T: Element + Copy>(arr: PyReadonlyArray2<T>) -> (Vec<T>, Vec
         z.push(row[2]);
     }
     (x, y, z)
+}
+
+// Copy a row-major (N,3) Python array to a vector of Vec3 for octree operations in Rust
+fn pyarray_to_vec3(arr: PyReadonlyArray2<f64>) -> Vec<Vec3> {
+    let n = arr.shape()[0];
+    let mut out: Vec<Vec3> = Vec::with_capacity(n);
+
+    for row in arr.as_array().rows() {
+        out.push(Vec3([row[0], row[1], row[2]]));
+    }
+    out
 }
 
 // Transpose 3 column vectors into a row-major (N,3) PyArray2
@@ -143,66 +156,83 @@ pub fn h_current_tet4_direct<'py>(
 }
 
 #[pyfunction]
-fn _hfield_tetrahedrons(
-    nodes_flat: PyReadonlyArray1<f64>,
-    centroids_flat: PyReadonlyArray1<f64>,
-    vol: PyReadonlyArray1<f64>,
-    jdensity_flat: PyReadonlyArray1<f64>,
-    x: PyReadonlyArray1<f64>,
-    y: PyReadonlyArray1<f64>,
-    z: PyReadonlyArray1<f64>,
-    mut bx: PyReadwriteArray1<f64>,
-    mut by: PyReadwriteArray1<f64>,
-    mut bz: PyReadwriteArray1<f64>,
+fn h_current_tet4_octree<'py>(
+    py: Python<'py>,
+    nodes: PyReadonlyArray2<f64>,
+    connectivity: PyReadonlyArray2<u32>,
+    jdensity: PyReadonlyArray2<f64>,
+    tgt_pts: PyReadonlyArray2<f64>,
     theta: f64,
     leaf_threshold: u32,
     nthreads_requested: u32,
-) -> PyResult<()> {
-    use crate::octree::{CurrentSources, HFieldSolver, Octree, tet_element};
+) -> PyResult<Bound<'py, PyArray2<f64>>> {
+    let n_srcs: usize = connectivity.shape()[0];
+    let n_tgts: usize = tgt_pts.shape()[0];
+    let max_depth: u8 = 21;
+    let mut centroids: Vec<Vec3> = vec![Vec3::default(); n_srcs];
+    let mut volumes: Vec<f64> = vec![0.0; n_srcs];
+    let _nodes = to_vec3s(nodes.as_slice()?);
+    let _connectivity = to_u32x4s(connectivity.as_slice()?);
+
+    let (x, y, z) = pyarray_to_3cols(tgt_pts);
+    let (mut hx, mut hy, mut hz) = col_buffer(n_tgts);
+
+    mesh::centroids(_nodes, _connectivity, &mut centroids);
+    mesh::volumes(_nodes, _connectivity, &mut volumes);
+
     let mut sources: CurrentSources<tet_element::TetSources> =
         CurrentSources(tet_element::TetSources::new(
-            nodes_flat.as_slice()?,
-            centroids_flat.as_slice()?,
-            vol.as_slice()?,
-            jdensity_flat.as_slice()?,
+            _nodes,
+            _connectivity,
+            &centroids,
+            &volumes,
+            to_vec3s(jdensity.as_slice()?),
         ));
-    let max_depth: u8 = 21;
 
     let tree: Octree<CurrentSources<tet_element::TetSources>> =
         Octree::build_from_sources(sources, max_depth, leaf_threshold);
 
     tree.h_field_parallel(
-        (x.as_slice()?, y.as_slice()?, z.as_slice()?),
-        (bx.as_slice_mut()?, by.as_slice_mut()?, bz.as_slice_mut()?),
+        (&x, &y, &z),
+        (&mut hx, &mut hy, &mut hz),
         theta,
         nthreads_requested,
     );
-    Ok(())
+    Ok(cols_to_pyarray(py, (hx, hy, hz)))
 }
 
 #[pyfunction]
-fn _hfield_dipole_tetrahedrons(
-    nodes_flat: PyReadonlyArray1<f64>,
-    centroids_flat: PyReadonlyArray1<f64>,
-    vol: PyReadonlyArray1<f64>,
-    jdensity_flat: PyReadonlyArray1<f64>,
-    x: PyReadonlyArray1<f64>,
-    y: PyReadonlyArray1<f64>,
-    z: PyReadonlyArray1<f64>,
-    mut bx: PyReadwriteArray1<f64>,
-    mut by: PyReadwriteArray1<f64>,
-    mut bz: PyReadwriteArray1<f64>,
+fn _hfield_dipole_tetrahedrons<'py>(
+    py: Python<'py>,
+    nodes: PyReadonlyArray2<f64>,
+    connectivity: PyReadonlyArray2<u32>,
+    magnetization: PyReadonlyArray2<f64>,
+    tgt_pts: PyReadonlyArray2<f64>,
     theta: f64,
     leaf_threshold: u32,
     nthreads_requested: u32,
-) -> PyResult<()> {
-    use crate::octree::{DipoleSources, HFieldSolver, Octree, tet_element};
+) -> PyResult<Bound<'py, PyArray2<f64>>> {
+    let n_srcs = connectivity.shape()[0];
+    let n_tgts: usize = tgt_pts.shape()[0];
+    let max_depth: u8 = 21;
+    let mut centroids: Vec<Vec3> = vec![Vec3::default(); n_srcs];
+    let mut volumes: Vec<f64> = vec![0.0; n_srcs];
+    let _nodes = to_vec3s(nodes.as_slice()?);
+    let _connectivity = to_u32x4s(connectivity.as_slice()?);
+
+    let (x, y, z) = pyarray_to_3cols(tgt_pts);
+    let (mut hx, mut hy, mut hz) = col_buffer(n_tgts);
+
+    mesh::centroids(_nodes, _connectivity, &mut centroids);
+    mesh::volumes(_nodes, _connectivity, &mut volumes);
+
     let mut sources: DipoleSources<tet_element::TetSources> =
         DipoleSources(tet_element::TetSources::new(
-            nodes_flat.as_slice()?,
-            centroids_flat.as_slice()?,
-            vol.as_slice()?,
-            jdensity_flat.as_slice()?,
+            _nodes,
+            _connectivity,
+            &centroids,
+            &volumes,
+            to_vec3s(magnetization.as_slice()?),
         ));
     let max_depth: u8 = 21;
 
@@ -210,12 +240,12 @@ fn _hfield_dipole_tetrahedrons(
         Octree::build_from_sources(sources, max_depth, leaf_threshold);
 
     tree.h_field_parallel(
-        (x.as_slice()?, y.as_slice()?, z.as_slice()?),
-        (bx.as_slice_mut()?, by.as_slice_mut()?, bz.as_slice_mut()?),
+        (&x, &y, &z),
+        (&mut hx, &mut hy, &mut hz),
         theta,
         nthreads_requested,
     );
-    Ok(())
+    Ok(cols_to_pyarray(py, (hx, hy, hz)))
 }
 
 #[pyfunction]
@@ -367,35 +397,38 @@ fn _h_demag_tet4(
 // ---
 
 #[pyfunction]
-fn _mesh_centroids(
-    nodes_flat: PyReadonlyArray1<f64>,
-    connectivity_flat: PyReadonlyArray1<u32>,
-    mut x: PyReadwriteArray1<f64>,
-    mut y: PyReadwriteArray1<f64>,
-    mut z: PyReadwriteArray1<f64>,
-) -> PyResult<()> {
+fn mesh_centroids<'py>(
+    py: Python<'py>,
+    nodes: PyReadonlyArray2<f64>,
+    connectivity: PyReadonlyArray2<u32>,
+) -> PyResult<Bound<'py, PyArray2<f64>>> {
+    let n_centroids: usize = connectivity.shape()[0];
+    let mut out: Vec<f64> = vec![0.0; n_centroids * 3];
+
     mesh::centroids(
-        nodes_flat.as_slice()?,
-        connectivity_flat.as_slice()?,
-        x.as_slice_mut()?,
-        y.as_slice_mut()?,
-        z.as_slice_mut()?,
+        to_vec3s(nodes.as_slice()?),
+        to_u32x4s(connectivity.as_slice()?),
+        to_vec3s_mut(&mut out),
     );
-    Ok(())
+
+    PyArray1::from_vec(py, out).reshape([n_centroids, 3])
 }
 
 #[pyfunction]
-fn _mesh_volumes(
-    nodes: PyReadonlyArray1<f64>,
-    connectivity: PyReadonlyArray1<u32>,
-    mut vol: PyReadwriteArray1<f64>,
-) -> PyResult<()> {
+fn mesh_volumes<'py>(
+    py: Python<'py>,
+    nodes: PyReadonlyArray2<f64>,
+    connectivity: PyReadonlyArray2<u32>,
+) -> PyResult<Bound<'py, PyArray1<f64>>> {
+    let n_elements = connectivity.shape()[0];
+
+    let mut out: Vec<f64> = vec![0.0; n_elements];
     mesh::volumes(
-        nodes.as_slice()?,
-        connectivity.as_slice()?,
-        vol.as_slice_mut()?,
+        to_vec3s(nodes.as_slice()?),
+        to_u32x4s(connectivity.as_slice()?),
+        &mut out,
     );
-    Ok(())
+    Ok(PyArray1::from_vec(py, out))
 }
 
 #[pyfunction]
@@ -482,15 +515,16 @@ fn _mesh_surface_tets<'py>(
 fn _oersted<'py>(_py: Python, m: Bound<'py, PyModule>) -> PyResult<()> {
     m.add_function(wrap_pyfunction!(b_current_point_direct, m.clone())?)?;
     m.add_function(wrap_pyfunction!(h_current_point_octree, m.clone())?)?;
-    m.add_function(wrap_pyfunction!(_hfield_tetrahedrons, m.clone())?)?;
+    m.add_function(wrap_pyfunction!(h_current_tet4_direct, m.clone())?)?;
+    m.add_function(wrap_pyfunction!(h_current_tet4_octree, m.clone())?)?;
     m.add_function(wrap_pyfunction!(_hfield_dipole, m.clone())?)?;
     m.add_function(wrap_pyfunction!(h_current_tet4_direct, m.clone())?)?;
     m.add_function(wrap_pyfunction!(_hfield_dipole_tetrahedrons, m.clone())?)?;
     m.add_function(wrap_pyfunction!(_h_demag_tet4, m.clone())?)?;
 
     // Mesh functions
-    m.add_function(wrap_pyfunction!(_mesh_centroids, m.clone())?)?;
-    m.add_function(wrap_pyfunction!(_mesh_volumes, m.clone())?)?;
+    m.add_function(wrap_pyfunction!(mesh_centroids, m.clone())?)?;
+    m.add_function(wrap_pyfunction!(mesh_volumes, m.clone())?)?;
     m.add_function(wrap_pyfunction!(_mesh_surface_faces, m.clone())?)?;
     m.add_function(wrap_pyfunction!(_mesh_surface_face_properties, m.clone())?)?;
     m.add_function(wrap_pyfunction!(_mesh_surface_forces, m.clone())?)?;